Implement Torsten's quorum enumeration algo:  - replace methods in quorum_lachowski with  - new module quorums containing `enumerate_quorums` and `traverse_quorums`
